add checking network connection

+++ lib/features/auth/domain/use_cases/change_email_use_case.dart
@@ -1,12 +1,19 @@
+import 'package:interview_master/core/utils/network_info.dart';
 import 'package:interview_master/features/auth/domain/repositories/auth_repository.dart';
+import '../../../../core/errors/network_exception.dart';
 import '../../data/models/my_user.dart';
 
 class ChangeEmailUseCase {
   final AuthRepository _authRepository;
+  final NetworkInfo _networkInfo;
 
-  ChangeEmailUseCase(this._authRepository);
+  ChangeEmailUseCase(this._authRepository, this._networkInfo);
 
   Future<void> call(String email, String password) async {
+    final isConnected = await _networkInfo.isConnected;
+    if (!isConnected) {
+      throw NetworkException();
+    }
     final user = await _authRepository.getUser();
     await _authRepository.deleteAccount();
     await _authRepository.signUp(
